core: move clock utilities to shared location
